+++ CHANGELOG.md
@@ -6,6 +6,7 @@ updates feed: https://xyrio.github.io/RSSOwlnix-site/updates.rss
 - updated httpclient to 5.5
 - removed auto reload of feeds without any data and error on viewing feed or the folder(s) feed is in [#118](https://github.com/Xyrio/RSSOwlnix/issues/118)
 - added RSSOWLNIX_USER_AGENT system property to change the User-Agent used when requesting feeds to match demands of websites for popular browsers [#164](https://github.com/Xyrio/RSSOwlnix/issues/164)
+- changed name reverse sorting to reverse only word positions so that numbers are sorted as usual
 
 # 2.9.0-beta
 - runs also with java 15, **does not run with java 16 [#116](https://github.com/Xyrio/RSSOwlnix/issues/116)**